move spring adapter into plugin
