+++ sofa-serverless-runtime/sofa-serverless-base-plugin/src/main/resources/META-INF/spring.factories
@@ -1,8 +1,8 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.alipay.sofa.serverless.plugin.BaseRuntimeAutoConfiguration
 org.springframework.context.ApplicationListener=\
-  com.alipay.sofa.serverless.common.spring.ServerlessApplicationListener
+  com.alipay.sofa.serverless.plugin.spring.ServerlessApplicationListener
 org.springframework.boot.env.EnvironmentPostProcessor=\
-  com.alipay.sofa.serverless.common.spring.ServerlessEnvironmentPostProcessor
+  com.alipay.sofa.serverless.plugin.spring.ServerlessEnvironmentPostProcessor
