+++ sofa-serverless-runtime/sofa-serverless-base-plugin/src/test/java/com/alipay/sofa/serverless/plugin/spring/ServerlessApplicationListenerTest.java
@@ -14,9 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.serverless.common;
+package com.alipay.sofa.serverless.plugin.spring;
 
-import com.alipay.sofa.serverless.common.spring.ServerlessApplicationListener;
 import org.junit.Test;
 import org.springframework.boot.ConfigurableBootstrapContext;
 import org.springframework.boot.SpringApplication;
