+++ sofa-serverless-runtime/sofa-serverless-common/pom.xml
@@ -47,10 +47,6 @@
             <artifactId>junit</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-        </dependency>
 
     </dependencies>
